[ refactor ] description부분 분리
